+++ Tracker/CoreData/Stores/DaysStore.swift
@@ -9,8 +9,12 @@ import UIKit
 import CoreData
 
 final class DaysStore {
+    // MARK: Properties
+    
     private let context: NSManagedObjectContext
     private let dataManager = CoreDataManager.shared
+    
+    // MARK: Init
     
     init(context: NSManagedObjectContext) {
         self.context = context
@@ -22,6 +26,8 @@ final class DaysStore {
         let context = CoreDataManager.shared.persistentContainer.viewContext
         self.init(context: context)
     }
+    
+    // MARK: Methods
     
     private func save() {
         dataManager.saveContext()
