+++ Tracker/TabBar/TabBarViewController.swift
@@ -8,11 +8,16 @@
 import UIKit
 
 final class TabBarViewController: UITabBarController {
+    
+    // MARK: Life Cycle
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         
         configure()
     }
+    
+    // MARK: Methods
     
     private func configure() {
         view.backgroundColor = Resources.Colors.white
@@ -40,21 +45,5 @@ final class TabBarViewController: UITabBarController {
         )
         
         self.viewControllers  = [trackersNav, statisticNav]
-    }
-}
-
-extension UITabBar {
-    func addTopBorder(color: UIColor?, thickness: CGFloat) {
-        let subview = UIView()
-        subview.translatesAutoresizingMaskIntoConstraints = false
-        subview.backgroundColor = color
-        self.addSubview(subview)
-        
-        NSLayoutConstraint.activate([
-            subview.leftAnchor.constraint(equalTo: self.leftAnchor),
-            subview.rightAnchor.constraint(equalTo: self.rightAnchor),
-            subview.heightAnchor.constraint(equalToConstant: thickness),
-            subview.topAnchor.constraint(equalTo: self.topAnchor)
-        ])
     }
 }
