+++ Tracker/TrackerCreation/Controllers/NewTrackerViewController.swift
@@ -9,6 +9,7 @@ import UIKit
 
 final class NewTrackerViewController: UIViewController {
     // MARK: Properties
+    
     weak var delegate: NewTrackerViewControllerDelegate?
     private let newHabitButton = BasicLargeButton(title: "Привычка")
     private let newEventButton = BasicLargeButton(title: "Нерегулярное событие")
@@ -21,8 +22,39 @@ final class NewTrackerViewController: UIViewController {
         configure()
     }
     
-    // MARK: Methods
+    // MARK: Actions
     
+    @objc private func newHabitButtonTapped() {
+        let habitViewController = NewHabitOrEventViewController(type: .habit)
+        habitViewController.delegate = self
+        let habitNav = UINavigationController(rootViewController: habitViewController)
+        present(habitNav, animated: true)
+    }
+    
+    @objc private func newEventButtonTapped() {
+        let eventViewController = NewHabitOrEventViewController(type: .event)
+        eventViewController.delegate = self
+        let eventNav = UINavigationController(rootViewController: eventViewController)
+        present(eventNav, animated: true)
+    }
+}
+
+// MARK: NewHabitOrEventViewControllerDelegate
+
+extension NewTrackerViewController: NewHabitOrEventViewControllerDelegate {
+    func closeController() {
+        self.dismiss(animated: true)
+    }
+    
+    func addTracker(tracker: Tracker, category: TrackerCategory) {
+        delegate?.addTracker(tracker: tracker, category: category)
+        self.dismiss(animated: true)
+    }
+}
+
+// MARK: UI
+
+extension NewTrackerViewController {
     private func configure() {
         view.backgroundColor = Resources.Colors.white
         newHabitButton.addTarget(self, action: #selector(newHabitButtonTapped), for: .touchUpInside)
@@ -57,32 +89,5 @@ final class NewTrackerViewController: UIViewController {
             newEventButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -20),
             newEventButton.heightAnchor.constraint(equalToConstant: 60)
         ])
-    }
-    
-    // MARK: Actions
-    
-    @objc private func newHabitButtonTapped() {
-        let habitViewController = NewHabitOrEventViewController(type: .habit)
-        habitViewController.delegate = self
-        let habitNav = UINavigationController(rootViewController: habitViewController)
-        present(habitNav, animated: true)
-    }
-    
-    @objc private func newEventButtonTapped() {
-        let eventViewController = NewHabitOrEventViewController(type: .event)
-        eventViewController.delegate = self
-        let eventNav = UINavigationController(rootViewController: eventViewController)
-        present(eventNav, animated: true)
-    }
-}
-
-extension NewTrackerViewController: NewHabitOrEventViewControllerDelegate {
-    func closeController() {
-        self.dismiss(animated: true)
-    }
-    
-    func addTracker(tracker: Tracker, category: TrackerCategory) {
-        delegate?.addTracker(tracker: tracker, category: category)
-        self.dismiss(animated: true)
     }
 }
