--- Tracker/TrackerCreation/Models/CollectionType.swift
@@ -1,12 +1,0 @@
-//
-//  CollectionType.swift
-//  Tracker
-//
-//  Created by Александр Плешаков on 30.05.2024.
-//
-
-import Foundation
-
-enum CollectionType {
-    case emoji, color
-}
